fix: Use new MasonryInfiniteScrollArtworkGrid for notification artwork list
