+++ src/app/Components/ArtworkGrids/MasonryArtworkGridItem.tsx
@@ -1,6 +1,6 @@
 import { ScreenOwnerType } from "@artsy/cohesion"
 import { Flex, useScreenDimensions, useSpace } from "@artsy/palette-mobile"
-import ArtworkGridItem from "app/Components/ArtworkGrids/ArtworkGridItem"
+import ArtworkGridItem, { PriceOfferMessage } from "app/Components/ArtworkGrids/ArtworkGridItem"
 import { useNavigateToPageableRoute } from "app/system/navigation/useNavigateToPageableRoute"
 import { NUM_COLUMNS_MASONRY } from "app/utils/masonryHelpers"
 import { ViewProps } from "react-native"
@@ -29,6 +29,7 @@ interface MasonryArtworkGridItemProps {
   navigateToPageableRoute: ReturnType<typeof useNavigateToPageableRoute>["navigateToPageableRoute"]
   numColumns?: number
   artworkMetaStyle?: ViewProps["style"]
+  priceOfferMessage?: PriceOfferMessage
 }
 
 export const MasonryArtworkGridItem: React.FC<MasonryArtworkGridItemProps> = ({
@@ -42,6 +43,7 @@ export const MasonryArtworkGridItem: React.FC<MasonryArtworkGridItemProps> = ({
   navigateToPageableRoute,
   numColumns = NUM_COLUMNS_MASONRY,
   artworkMetaStyle = {},
+  priceOfferMessage,
   ...rest
 }) => {
   const space = useSpace()
@@ -65,6 +67,7 @@ export const MasonryArtworkGridItem: React.FC<MasonryArtworkGridItemProps> = ({
         height={imgHeight}
         navigateToPageableRoute={navigateToPageableRoute}
         artworkMetaStyle={artworkMetaStyle}
+        priceOfferMessage={priceOfferMessage}
       />
     </Flex>
   )
