+++ src/app/Components/ArtworkRail/ArtworkRailCard.tsx
@@ -45,7 +45,6 @@ export const ARTWORK_RAIL_CARD_IMAGE_HEIGHT = {
 const ARTWORK_LARGE_RAIL_CARD_IMAGE_WIDTH = 295
 
 export type ArtworkCardSize = "small" | "large" | "extraLarge" | "fullWidth"
-export type PriceOfferMessage = { priceListedMessage: string; priceWithDiscountMessage: string }
 
 export interface ArtworkRailCardProps extends ArtworkActionTrackingProps {
   artwork: ArtworkRailCard_artwork$key
@@ -63,7 +62,6 @@ export interface ArtworkRailCardProps extends ArtworkActionTrackingProps {
   showSaveIcon?: boolean
   size: ArtworkCardSize
   testID?: string
-  priceOfferMessage?: PriceOfferMessage
 }
 
 export const ArtworkRailCard: React.FC<ArtworkRailCardProps> = ({
@@ -81,7 +79,6 @@ export const ArtworkRailCard: React.FC<ArtworkRailCardProps> = ({
   showSaveIcon = false,
   size,
   testID,
-  priceOfferMessage,
   ...restProps
 }) => {
   const EXTRALARGE_RAIL_CARD_IMAGE_WIDTH = useExtraLargeWidth()
@@ -195,11 +192,6 @@ export const ArtworkRailCard: React.FC<ArtworkRailCardProps> = ({
   const displayForRecentlySoldArtwork =
     !!isRecentlySoldArtwork && (size === "large" || size === "extraLarge")
 
-  const displayPriceOfferMessage =
-    !!priceOfferMessage &&
-    !!priceOfferMessage.priceListedMessage &&
-    !!priceOfferMessage.priceWithDiscountMessage
-
   return (
     <AnalyticsContextProvider
       contextScreenOwnerId={contextScreenOwnerId}
@@ -299,24 +291,6 @@ export const ArtworkRailCard: React.FC<ArtworkRailCardProps> = ({
                   </Text>
                 )}
 
-                {!!displayPriceOfferMessage && (
-                  <Flex flexDirection="row">
-                    <Text
-                      lineHeight="20px"
-                      variant="xs"
-                      color={primaryTextColor}
-                      numberOfLines={1}
-                      fontWeight="bold"
-                    >
-                      {priceOfferMessage.priceWithDiscountMessage}
-                    </Text>
-                    <Text color="black60" variant="xs">
-                      {" "}
-                      (List price: {priceOfferMessage.priceListedMessage})
-                    </Text>
-                  </Flex>
-                )}
-
                 {!!isRecentlySoldArtwork && (size === "large" || size === "extraLarge") && (
                   <RecentlySoldCardSection
                     priceRealizedDisplay={priceRealizedDisplay}
@@ -327,7 +301,7 @@ export const ArtworkRailCard: React.FC<ArtworkRailCardProps> = ({
                   />
                 )}
 
-                {!!saleMessage && !isRecentlySoldArtwork && !displayPriceOfferMessage && (
+                {!!saleMessage && !isRecentlySoldArtwork && (
                   <Text
                     lineHeight="20px"
                     variant="xs"
